Include previous values in UPDATE and DELETE ops
In order to write processors that monitors for certain state changes it
can use the previous values to compare a transition.

This PR adds the field `previous_data` to UPDATE and DELETE statements
(it'll add an empty json object for INSERT statements).
I'm not sure about the name of this field. Any other suggestions?

Also had to update some gometalinter issues.

Closes #27